requirements.txt
Django==5.0.3
Pillow==10.2.0
environs==11.0.0
djangorestframework==3.14.0
stripe==8.6.0
django_phonenumber_field==7.3.0
phonenumbers==8.13.32
Babel-2.14
